+++ Application/app/components/RangeSliderInput.tsx
